+++ patch/main.js
@@ -3,7 +3,7 @@
  * This code is released under GNU General Public License (https://github.com/alcapitan/atui/blob/dev/LICENSE.md)
  */
 
-atuiKernel_FooterLastedited(5, 4, 2023);
+atuiKernel_FooterLastedited(6, 4, 2023);
 atuiKernel_ColorschemeGeneratorAuto([230, 51, 0]);
 
 atuiKernel_ToolsContextmenu("atuiKernel_NavigationImg");
@@ -27,8 +27,9 @@ atuiKernel_ToolsInfotip("atuiKernel_SectionDownloadChecksum");*/
 /*atuiKernel_NotificationDisplay('alert','default',["console.log('ok')"],["console.log('annuler')"],'Maecenas dictum felis tristique lectus imperdiet aliquet. ','Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris condimentum, turpis vel efficitur dictum, ipsum tortor pharetra velit, at mollis nisl ligula pellentesque mauris. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. ');*/
 /*atuiKernel_NotificationCookies();*/
 
-document.getElementById("clipboardButton").addEventListener("click", function () {
-    text = document.getElementById("clipboardInput").value;
+document.getElementById("clipboardButton").addEventListener("click", async function () {
+    const text = document.getElementById("clipboardInput").value;
     atuiKernel_ClipboardCopy(text);
-    document.getElementById("clipboard-paste").innerText = atuiKernel_ClipboardPaste();
+    const clipboardText = await atuiKernel_ClipboardPaste();
+    document.getElementById("clipboard-paste").innerText = clipboardText;
 });